f
Working up to a point. Now determines the panes in an active col/row
and only changes them.

This still uses the initial settings stage to
determine how many splits/panes are in each row/col (taking max val).
This therefore means that it does not always correctly resize a col/row
if another col/row has more or less than the active.